+++ test/mmdb2_decoder/pointer_test.exs
@@ -14,4 +14,16 @@ defmodule MMDB2Decoder.PointerTest do
 
     assert {:ok, result_bang} == result_regular
   end
+
+  test "lookup_pointer!/3 equals lookup_pointer/3" do
+    {:ok, _, _, data} =
+      :fixture_ipv4_24
+      |> Fixture.contents()
+      |> MMDB2Decoder.parse_database()
+
+    result_bang = MMDB2Decoder.lookup_pointer!(0, data)
+    result_regular = MMDB2Decoder.lookup_pointer(0, data)
+
+    assert {:ok, result_bang} == result_regular
+  end
 end